+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - thermostat accessory: show target slider when available
 - updated icon for door, window and window covering services
 - revert back to font-awesome v6 for compatibility
+- accessory info: show characteristic props with chevrons
 
 ### Other Changes